glTF: upgrade Draco to version 1.3.5 and add mesh skinning support
This will fix exporting meshes with armatures using Draco compression, like:
https://github.com/KhronosGroup/glTF-Blender-IO/issues/617

Differential Revision: https://developer.blender.org/D6342

+++ extern/draco/dracoenc/src/draco/io/parser_utils.cc
@@ -128,7 +128,7 @@ bool ParseFloat(DecoderBuffer *buffer, float *value) {
         return false;
 
       // Apply exponent scaling to value.
-      v *= pow(10.0, exponent);
+      v *= pow(static_cast<double>(10.0), exponent);
     }
   }
